+++ akshare/bond/bond_convert.py
@@ -1,44 +1,27 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 """
-Date: 2022/6/20 17:05
+Date: 2022/9/24 15:05
 Desc: 债券-集思录-可转债
 集思录：https://app.jisilu.cn/data/cbnew/#cb
 """
-import ast
-
 import pandas as pd
 import requests
-from bs4 import BeautifulSoup
 
 from akshare.utils import demjson
 
 
 def bond_cb_index_jsl() -> pd.DataFrame:
     """
-    https://www.jisilu.cn/data/cbnew/cb_index/
+    https://www.jisilu.cn/web/data/cb/index
     首页-可转债-集思录可转债等权指数
     :return: 集思录可转债等权指数
     :rtype: pandas.DataFrame
     """
-    url = "https://www.jisilu.cn/data/cbnew/cb_index/"
+    url = "https://www.jisilu.cn/webapi/cb/index_history/"
     r = requests.get(url)
-    soup = BeautifulSoup(r.text, "lxml")
-    data_text = soup.find_all("script", attrs={"type": "text/javascript"})[
-        -4
-    ].string
-    inner_data_text = data_text[
-        data_text.find("__date") : data_text.find("__data")
-    ].strip("__date = ")
-    date_list = ast.literal_eval(inner_data_text[: inner_data_text.rfind(";")])
-    inner_data_text = data_text[
-        data_text.find("__data") : data_text.find("for(var")
-    ].strip("__data = ")
-    data_dict = demjson.decode(inner_data_text[: inner_data_text.rfind(";")])
-    temp_df = pd.DataFrame([date_list, data_dict["price"]]).T
-    temp_df.columns = ["date", "price"]
-    temp_df["date"] = pd.to_datetime(temp_df["date"]).dt.date
-    temp_df["price"] = pd.to_numeric(temp_df["price"])
+    data_dict = demjson.decode(r.text)["data"]
+    temp_df = pd.DataFrame(data_dict)
     return temp_df
 
 
@@ -294,7 +277,7 @@ def bond_cb_redeem_jsl() -> pd.DataFrame:
             "强赎价",
             "强赎天计数",
             "强赎条款",
-            '强赎状态'
+            "强赎状态",
         ]
     ]
     temp_df["现价"] = pd.to_numeric(temp_df["现价"])
@@ -306,8 +289,12 @@ def bond_cb_redeem_jsl() -> pd.DataFrame:
     temp_df["强赎触发价"] = pd.to_numeric(temp_df["强赎触发价"])
     temp_df["正股价"] = pd.to_numeric(temp_df["正股价"])
     temp_df["强赎价"] = pd.to_numeric(temp_df["强赎价"], errors="coerce")
-    temp_df["强赎天计数"] = temp_df["强赎天计数"].replace(r'^.*?(\d{1,2}\/\d{1,2} \| \d{1,2}).*?$',r'\1', regex=True)
-    temp_df["强赎状态"] = temp_df["强赎状态"].map({'R':'已公告强赎','O':'公告要强赎','G':'公告不强赎','B':'已满足强赎条件','':''})
+    temp_df["强赎天计数"] = temp_df["强赎天计数"].replace(
+        r"^.*?(\d{1,2}\/\d{1,2} \| \d{1,2}).*?$", r"\1", regex=True
+    )
+    temp_df["强赎状态"] = temp_df["强赎状态"].map(
+        {"R": "已公告强赎", "O": "公告要强赎", "G": "公告不强赎", "B": "已满足强赎条件", "": ""}
+    )
     return temp_df